requirements.txt
absl-py==1.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
asgiref==3.4.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==2.0.0
backcall==0.2.0
backoff==2.2.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.5
comm==0.1.3
commonmark==0.9.1
cryptography==40.0.1
cycler==0.11.0
Cython==0.29.34
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.9
dlib @ file:///D:/Python%20Programming/face_recog_dlib_file-master/face_recog_dlib_file-master/dlib-19.22.99-cp310-cp310-win_amd64.whl
dnspython==2.2.1
et-xmlfile==1.1.0
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.10.7
Flask==2.2.2
fonttools==4.31.1
frozenlist==1.3.3
greenlet==2.0.2
grpcio==1.51.1
grpcio-tools==1.51.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.0
langchain @ git+https://github.com/hwchase17/langchain.git@955bd2e1db8d008d628963cb8d2bad5c1d354744
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mediapipe==0.8.9.1
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numpy==1.22.3
olefile==0.46
openai==0.25.0
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.5.3
pandas-stubs==1.5.2.221213
parso==0.8.3
pdfminer==20191125
pdfminer.six==20201018
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.21.12
protobuf2pydantic==2021.11.25
psutil==5.9.4
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
PyGetWindow==0.0.9
Pygments==2.15.0
PyJWT==2.6.0
pymongo==4.3.3
PyMsgBox==1.0.9
pypandoc==1.11
pyparsing==3.0.7
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-magic==0.4.27
python-pptx==0.6.21
pytweening==1.0.4
pytz==2022.6
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.1
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.9.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.47
sqlparse==0.4.2
stack-data==0.6.2
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.3
twilio==7.16.0
typer==0.7.0
types-pytz==2022.6.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
unstructured==0.5.11
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.11.0